[Refactor]: 수익률 계산 로직을 ReturnRateCalculator 클래스로 분리
