+++ __tests__/LottoWinnerVerifierTest.js
@@ -18,18 +18,4 @@ describe('LottoWinnerVerifier 클래스 테스트', () => {
     };
     expect(outcome).toEqual(result);
   });
-
-  test('로또 수익률 테스트', () => {
-    const outcome = {
-      first: 0,
-      second: 0,
-      third: 0,
-      fourth: 0,
-      fifth: 1,
-      ['no prize']: 0,
-    };
-    expect(LottoWinnerVerifier.calulateLottoPayoutRate(outcome, 8000)).toBe(
-      62.5
-    );
-  });
 });
