+++ src/constant/index.js
@@ -1,9 +1,9 @@
 import { BONUS_NUMBER, BUYING_PRICE, LOTTO_NUMBERS } from './error.js';
-import { NUMBERS, PRICE } from './lotto.js';
+import { NUMBERS, PRICE, PRIZE } from './lotto.js';
 import INPUT from './input.js';
 import RANDOM from './random.js';
 
 const ERROR = { BONUS_NUMBER, BUYING_PRICE, LOTTO_NUMBERS };
-const LOTTO = { NUMBERS, PRICE };
+const LOTTO = { NUMBERS, PRICE, PRIZE };
 
 export { ERROR, INPUT, LOTTO, RANDOM };
