+++ src/constant/lotto.js
@@ -4,5 +4,11 @@ const NUMBERS = {
   MAX: 45,
 };
 const PRICE = 1000;
-
-export { NUMBERS, PRICE };
+const PRIZE = {
+  first: 2000000000,
+  second: 30000000,
+  third: 1500000,
+  fourth: 50000,
+  fifth: 5000,
+};
+export { NUMBERS, PRICE, PRIZE };
